# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Laravel Generator APIs",
        "version": "1.0.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/notifications/{id}": {
            "get": {
                "tags": [
                    "Notification"
                ],
                "summary": "Display the specified Notification",
                "description": "Get Notification",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Notification",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Notification"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Notification"
                ],
                "summary": "Update the specified Notification in storage",
                "description": "Update Notification",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Notification",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Notification that should be updated",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Notification"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Notification"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Notification"
                ],
                "summary": "Remove the specified Notification from storage",
                "description": "Delete Notification",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Notification",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Notification": {
            "required": [
                ""
            ],
            "properties": {
                "id": {
                    "description": "id",
                    "type": "string"
                },
                "title": {
                    "description": "title",
                    "type": "string"
                },
                "brief_description": {
                    "description": "brief_description",
                    "type": "string"
                },
                "type": {
                    "description": "type",
                    "type": "string"
                },
                "notifiable_type": {
                    "description": "notifiable_type",
                    "type": "string"
                },
                "notifiable_id": {
                    "description": "notifiable_id",
                    "type": "integer",
                    "format": "int32"
                },
                "notifier_type": {
                    "description": "notifier_type",
                    "type": "string"
                },
                "notifier_id": {
                    "description": "notifier_id",
                    "type": "integer",
                    "format": "int32"
                },
                "deadline": {
                    "description": "deadline",
                    "type": "string",
                    "format": "date-time"
                },
                "data": {
                    "description": "data",
                    "type": "string"
                },
                "read_at": {
                    "description": "read_at",
                    "type": "string",
                    "format": "date-time"
                },
                "created_at": {
                    "description": "created_at",
                    "type": "string",
                    "format": "date-time"
                },
                "updated_at": {
                    "description": "updated_at",
                    "type": "string",
                    "format": "date-time"
                },
                "deleted_at": {
                    "description": "deleted_at",
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
...
    }
}